+++ src/pages/_app.tsx
@@ -1,6 +1,7 @@
 import type { AppProps } from "next/app";
 import { ReactElement } from "react";
 import { Inter } from "@next/font/google";
+import { useRouter } from "next/router";
 import Navbar from "@/components/navbar/navbar";
 import "../styles/globals.scss";
 
@@ -10,10 +11,27 @@ const inter = Inter({
   variable: "--font-inter",
 });
 
+const navLinks = [
+  { name: "About", path: "about" },
+  { name: "Projects", path: "projects" },
+  { name: "Get Involved", path: "get-involved" },
+];
+
+function getTheme(path: string) {
+  if (path === "/about" || path === "/projects" || path === "/get-involved") {
+    return "dark";
+  } else {
+    return "light";
+  }
+}
+
 export default function App({ Component, pageProps }: AppProps): ReactElement {
+  const router = useRouter();
+  const theme = getTheme(router.pathname);
+
   return (
-    <div className={`${inter.variable}`}>
-      <Navbar />
+    <div className={`${inter.variable}`} style={{ position: "relative" }}>
+      <Navbar nav={navLinks} theme={theme} />
       <Component {...pageProps} />
     </div>
   );
