+++ src/styles/navbar/navbar.module.scss
@@ -4,6 +4,7 @@
   width: 100%;
   height: var(--navbar-height);
   position: absolute;
+  top: 0;
   display: flex;
   justify-content: center;
   align-items: center;
@@ -109,7 +110,7 @@
           > li {
             width: 100%;
             margin-left: calc(var(--edge-spacer) * 0.1);
-            transition: transform 0.8s ease-in-out 0.1s;
+            transition: transform 0.8s ease-in-out 0.2s;
             transform: translateX(-100%);
 
             > a {
